requirements.txt
amqp==5.3.1
asgiref==3.11.0
beautifulsoup4==4.14.3
billiard==4.2.4
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron_descriptor==2.0.6
curl_cffi==0.13.0
Django==5.2.10
django-celery-beat==2.8.1
django-timezone-field==7.2.1
djangorestframework==3.16.1
finnhub-python==2.4.27
frozendict==2.4.7
idna==3.11
kombu==5.6.2
multitasking==0.0.12
numpy==2.4.1
packaging==26.0
pandas==3.0.0
peewee==3.19.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
protobuf==6.33.4
pycparser==3.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
redis==6.4.0
requests==2.32.5
six==1.17.0
soupsieve==2.8.3
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.5.0
websockets==16.0
